+++ src/components/AssetExample.js
