+++ src/screens/MapScreen.js
@@ -1,0 +1,11 @@
+import react from "react";
+
+export default function MapScreen() {
+return(
+    <View>
+        <Text>
+            Map Screen
+        </Text>
+    </View>
+)
+}
